rename card models
