+++ lib/widgets/cards/card_widget.dart
@@ -1,8 +1,15 @@
 import 'package:cards/models/card_model.dart';
-import 'package:cards/widgets/cards/card_face_widget.dart';
-import 'package:cards/widgets/cards/skyjo_card_face_widget.dart';
+import 'package:cards/widgets/cards/card_face_french_widget.dart';
+import 'package:cards/widgets/cards/card_face_skyjo_widget.dart';
 import 'package:cards/widgets/wiggle_widget.dart';
 import 'package:flutter/material.dart';
+
+class CardDimensions {
+  static const double width = 100.0;
+  static const double height = 150.0;
+  static const double margin = 4.0;
+  static const double borderRadius = 4.0;
+}
 
 /// A widget that displays a playing card or a placeholder.
 ///
@@ -56,8 +63,8 @@ class CardWidget extends StatelessWidget {
             border: Border.all(color: Colors.black, width: 1),
           ),
           child: card.suit == ''
-              ? SkyjoCardFaceWidget(card: card)
-              : CardFaceWidget(card: card),
+              ? CardFaceSkyjoWidget(card: card)
+              : CardFaceFrenchWidget(card: card),
         ),
       ),
     );
